+++ src/main/java/com/dansup/server/api/auth/service/AuthService.java
@@ -1,66 +1,24 @@
 package com.dansup.server.api.auth.service;
 
-import lombok.extern.slf4j.Slf4j;
-import net.minidev.json.JSONObject;
-import net.minidev.json.parser.JSONParser;
-import net.minidev.json.parser.ParseException;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.ResponseEntity;
+import com.dansup.server.api.auth.dto.request.SignUpDto;
+import com.dansup.server.api.profile.domain.Profile;
+import com.dansup.server.api.user.domain.User;
+import com.dansup.server.api.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
-import org.springframework.web.client.RestTemplate;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
-@Slf4j
+@Transactional
+@RequiredArgsConstructor
 public class AuthService {
 
-    @Value("${spring.security.oauth2.client.registration.kakao.client-id}")
-    private String kakaoClientId;
-    @Value("${spring.security.oauth2.client.registration.kakao.redirect-uri}")
-    private String redirectUri;
-    @Value("${spring.security.oauth2.client.registration.kakao.client-secret}")
-    private String clientSecret;
+    private UserRepository userRepository;
 
-    public String getKakaoAccessToken(String code) throws ParseException {
+    public void signin(User user, SignUpDto signUpDto) {
 
-        MultiValueMap<String, String> requestBody = new LinkedMultiValueMap<>();
-        requestBody.add("grant_type", "authorization_code");
-        requestBody.add("client_id", kakaoClientId);
-        requestBody.add("redirect_uri", redirectUri);
-        requestBody.add("code", code);
-        requestBody.add("client_secret", clientSecret);
-
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("Content-type", "application/x-www-form-urlencoded;charset=utf-8");
-
-        HttpEntity<MultiValueMap<String, String>> entity = new HttpEntity<>(requestBody, headers);
-
-        RestTemplate rt = new RestTemplate();
-
-        ResponseEntity<String> response = rt.exchange(
-                "https://kauth.kakao.com/oauth/token",
-                HttpMethod.POST,
-                entity,
-                String.class
-        );
-
-        String result = response.getBody();
-        log.info("result: " + result);
-
-        JSONParser parser = new JSONParser();
-        Object obj = parser.parse(result);
-        JSONObject jsonObj = (JSONObject) obj;
-
-        String accessToken = jsonObj.get("access_token").toString();
-        log.info("access_token: " + accessToken);
-
-        return accessToken;
+        Profile profile = Profile.createProfile(signUpDto);
 
     }
-
 
 }
